requirements.txt
python-digitalocean
netifaces
service
